debugger: document markChanged() better
Remove the now-wrong markChanged() documentation added in
0ba7895f9d7db3e14a29a0e6686558366c4a5548: markChanged() is emitted when
a mark is added or removed at a document border, not when source code
changes.

2558fe0606da7ecf6d3ca24685ee2b4f46f672a1 reimplemented
BreakpointModel::markChanged() with an explanation in the commit
message. Update and transfer that explanation into comments, because it
is not obvious why an existing breakpoint has to be removed on
MarkAdded action.

f27e8c140a516b88d9617d9e49b0ab6e2adb8360 introduced the current code to
handle the mark context menu with markContextMenuRequested(). The
context menu interacts with markChanged() such that it can trigger
MarkAdded event, but never MarkRemoved events.

A notable known bug is that BreakpointModel::markChanged() conflicts
with the document reload operation and is unintentionally invoked
during it. The end result is either the incorrect removal of or buggy
addition of many breakpoint instances.